requirements.txt
yt-dlp==2025.12.8
rich
questionary
fastapi
uvicorn
jinja2
python-multipart
pystray
pillow